fix: check if remote actor should be able to delete the note or user or not

+++ packages/backend/src/services/ap/inbox/DeleteProcessor.ts
@@ -3,16 +3,19 @@ import NoteService from '../../NoteService.js';
 import UserService from '../../UserService.js';
 
 class DeleteProcessor {
-	private async findAndDelete(apId: string): Promise<boolean> {
+	private async findAndDelete(as: string, apId: string): Promise<boolean> {
+		let actor = await UserService.get({ apId: as });
+		if (!actor) return false;
+		
 		let user = await UserService.get({ apId: apId });
-		if (user) {
+		if (user && actor.host === user.host) {
 			logger.debug('delete', 'deleting user');
 			await UserService.delete({ apId: apId });
 			return true;
 		}
 
 		let note = await NoteService.get({ apId: apId });
-		if (note) {
+		if (note && note.user && note.user.host === actor.host) {
 			logger.debug('delete', 'deleting note');
 			await NoteService.delete({ apId: apId });
 			return true;
@@ -26,9 +29,9 @@ class DeleteProcessor {
 		if (!body.object) return false;
 
 		if (!body.object.type) {
-			return await this.findAndDelete(body.object);
+			return await this.findAndDelete(body.actor, body.object);
 		} else {
-			return await this.findAndDelete(body.object.id);
+			return await this.findAndDelete(body.actor, body.object.id);
 		}
 	}
 }
